+++ BugsnagPerformance/Assets/BugsnagPerformance/Scripts/Internal/SpanFactory.cs
@@ -11,7 +11,7 @@ namespace BugsnagUnityPerformance
     {
 
         [ThreadStatic]
-        private static Stack<ISpanContext> _contextStack;
+        private static Stack<WeakReference<ISpanContext>> _contextStack;
 
         private RNGCryptoServiceProvider _rNGCryptoServiceProvider = new RNGCryptoServiceProvider();
 
@@ -66,7 +66,7 @@ namespace BugsnagUnityPerformance
                 spanOptions.IsFirstClass = true;
             }
             var span = CreateSpan(name, SpanKind.SPAN_KIND_INTERNAL, spanOptions);
-            span.SetAttribute("bugsnag.span.category", "custom");
+            span.SetAttributeInternal("bugsnag.span.category", "custom");
             return span;
         }
 
@@ -94,18 +94,17 @@ namespace BugsnagUnityPerformance
                     traceId = GetNewTraceId();
                 }
             }
-           
-            var newSpan = new Span(name, kind, spanId, traceId, parentSpanId, spanOptions.StartTime, spanOptions.IsFirstClass, _onSpanEnd);
 
+            var newSpan = new Span(name, kind, spanId, traceId, parentSpanId, spanOptions.StartTime, spanOptions.IsFirstClass, _onSpanEnd);
             if (spanOptions.MakeCurrentContext)
             {
                 AddToContextStack(newSpan);
             }
-            newSpan.SetAttribute("net.host.connection.type",_currentConnectionType);
+            newSpan.SetAttributeInternal("net.host.connection.type", _currentConnectionType);
             return newSpan;
         }
 
-        internal Span CreateAutomaticNetworkSpan(BugsnagUnityWebRequest request,string url)
+        internal Span CreateAutomaticNetworkSpan(BugsnagUnityWebRequest request, string url)
         {
             var verb = request.method.ToUpper();
 
@@ -114,9 +113,9 @@ namespace BugsnagUnityPerformance
             var spanOptions = new SpanOptions { MakeCurrentContext = false };
 
             var span = CreateSpan("HTTP/" + verb, SpanKind.SPAN_KIND_CLIENT, spanOptions);
-            span.SetAttribute("bugsnag.span.category", "network");
-            span.SetAttribute("http.url", url);
-            span.SetAttribute("http.method", verb);
+            span.SetAttributeInternal("bugsnag.span.category", "network");
+            span.SetAttributeInternal("http.url", url);
+            span.SetAttributeInternal("http.method", verb);
             return span;
         }
 
@@ -134,9 +133,9 @@ namespace BugsnagUnityPerformance
                 options = new SpanOptions { MakeCurrentContext = false };
             }
             var span = CreateSpan("HTTP/" + httpVerb, SpanKind.SPAN_KIND_CLIENT, options);
-            span.SetAttribute("bugsnag.span.category", "network");
-            span.SetAttribute("http.url", url);
-            span.SetAttribute("http.method", httpVerb.ToString());
+            span.SetAttributeInternal("bugsnag.span.category", "network");
+            span.SetAttributeInternal("http.url", url);
+            span.SetAttributeInternal("http.method", httpVerb.ToString());
             return span;
         }
 
@@ -181,9 +180,9 @@ namespace BugsnagUnityPerformance
                 options = new SpanOptions { IsFirstClass = true };
             }
             var span = CreateSpan("[ViewLoad/UnityScene]" + sceneName, SpanKind.SPAN_KIND_INTERNAL, options);
-            span.SetAttribute("bugsnag.span.category", "view_load");
-            span.SetAttribute("bugsnag.view.type", "UnityScene");
-            span.SetAttribute("bugsnag.view.name", sceneName);            
+            span.SetAttributeInternal("bugsnag.span.category", "view_load");
+            span.SetAttributeInternal("bugsnag.view.type", "UnityScene");
+            span.SetAttributeInternal("bugsnag.view.name", sceneName);   
             return span;
         }
 
@@ -196,34 +195,48 @@ namespace BugsnagUnityPerformance
 
             while (_contextStack.Count > 0)
             {
-                var top = (Span)_contextStack.Peek();
-                if (top.Ended)
+                var top = _contextStack.Peek();
+                if(top == null)
                 {
                     _contextStack.Pop();
+                    continue;
+                }
+                if (top.TryGetTarget(out var spanContext))
+                {
+                    if (((Span)spanContext).Ended)
+                    {
+                        _contextStack.Pop();
+                    }
+                    else
+                    {
+                        return spanContext;
+                    }
                 }
                 else
                 {
-                    return top;
+                    _contextStack.Pop();
+                    continue;
                 }
             }
 
             return null;
         }
 
+
         private void AddToContextStack(ISpanContext spanContext)
         {
             if (_contextStack == null)
             {
-                _contextStack = new Stack<ISpanContext>();
+                _contextStack = new Stack<WeakReference<ISpanContext>>();
             }
-            _contextStack.Push(spanContext);
+            _contextStack.Push(new WeakReference<ISpanContext>(spanContext));
         }
 
         internal Span CreateAutoAppStartSpan(string name, string category)
         {
-            var span = CreateSpan(name, SpanKind.SPAN_KIND_CLIENT,new SpanOptions());
-            span.SetAttribute("bugsnag.span.category", category);
-            span.SetAttribute("bugsnag.app_start.type", "UnityRuntime");
+            var span = CreateSpan(name, SpanKind.SPAN_KIND_CLIENT, new SpanOptions());
+            span.SetAttributeInternal("bugsnag.span.category", category);
+            span.SetAttributeInternal("bugsnag.app_start.type", "UnityRuntime");
             span.IsAppStartSpan = true;
             return span;
         }