+++ features/fixtures/mazerunner/Assets/Scripts/Scenarios/Attributes/AddAttributesInCallbacks.cs
@@ -1,20 +1,30 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using BugsnagUnityPerformance;
 using UnityEngine;
 
-public class ManualNetworkSpan : Scenario
+public class AddAttributesInCallbacks : Scenario
 {
     public override void PreparePerformanceConfig(string apiKey, string host)
     {
         base.PreparePerformanceConfig(apiKey, host);
         SetMaxBatchSize(1);
+        Configuration.AddOnSpanEnd(MyConfigCallback);
+    }
+
+    private bool MyConfigCallback(Span span)
+    {
+        span.SetAttribute("config-callback", true);
+        return true;
     }
 
     public override void Run()
     {
         base.Run();
-        BugsnagPerformance.StartNetworkSpan(Main.MazeHost, HttpVerb.PATCH).EndNetworkSpan(202,123,321);;
+        var span = BugsnagPerformance.StartSpan("AddAttributesInCallbacks");
+        span.End();
     }
 
 }
+