+++ features/fixtures/mazerunner/Assets/Scripts/Scenarios/Configuration/CustomServiceName.cs
@@ -1,23 +1,21 @@
-﻿using System.Collections;
+using System.Collections;
 using System.Collections.Generic;
 using BugsnagUnityPerformance;
 using UnityEngine;
 
-public class EnabledReleaseStages : Scenario
+public class CustomServiceName : Scenario
 {
-
     public override void PreparePerformanceConfig(string apiKey, string host)
     {
         base.PreparePerformanceConfig(apiKey, host);
-        Configuration.ReleaseStage = "EnabledReleaseStages";
-        Configuration.EnabledReleaseStages = new[] { "EnabledReleaseStages" };
         SetMaxBatchSize(1);
+        Configuration.ServiceName = "custom.service.name";
     }
 
     public override void Run()
     {
-        base.Run();
-        BugsnagPerformance.StartSpan("EnabledReleaseStages").End();
+        var span = BugsnagPerformance.StartSpan("CustomServiceName");
+        span.End();
     }
 
 }