+++ features/fixtures/mazerunner/Assets/Scripts/Scenarios/Spans/ConfiguredSamplingRate1.cs
@@ -8,7 +8,7 @@ public class ConfiguredSamplingRate1 : Scenario
     public override void PreparePerformanceConfig(string apiKey, string host)
     {
         base.PreparePerformanceConfig(apiKey, host);
-        SetMaxBatchAgeSeconds(1);
+        SetMaxBatchSize(4);
     }
 
     public override void Run()